+++ src/de/mod.rs
@@ -3,7 +3,7 @@
 use crate::error::{ErrorKind, Result};
 use crate::FitDataRecord;
 use nom::number::complete::le_u16;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::io::Read;
 use std::rc::Rc;
 use log::warn;
@@ -14,6 +14,15 @@ mod decode;
 use decode::Decoder;
 mod parser;
 pub use parser::{FitDataMessage, FitDefinitionMessage, FitFileHeader};
+
+/// Decoding options for the deserializer
+#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, Eq, Ord, Hash)]
+pub enum DecodeOption {
+    /// Ignore header section checksum value if it exists
+    SkipHeaderCrcValidation,
+    /// Ignore data section checksum value
+    SkipDataCrcValidation,
+}
 
 /// Stores a FIT file object (header, message or CRC)
 #[derive(Clone, Debug)]
@@ -30,6 +39,8 @@ pub enum FitObject {
 
 /// Manages the deserialization of a FIT data stream into Rust constructs.
 struct Deserializer {
+    /// The current set of deserialization options
+    options: HashSet<DecodeOption>,
     /// Track the current set of FIT message definitions, these are what allows the format to
     /// be self describing.
     definitions: HashMap<u8, Rc<parser::FitDefinitionMessage>>,
@@ -47,11 +58,17 @@ impl Deserializer {
     /// Create the deserializer with an empty state
     fn new() -> Self {
         Deserializer {
+            options: HashSet::new(),
             definitions: HashMap::new(),
             position: 0,
             end_of_messages: 0,
             crc: 0,
         }
+    }
+
+    /// Add or remove decoding options for deserializer by manipulating the set
+    fn options_mut(&mut self) -> &mut HashSet<DecodeOption> {
+        &mut self.options
     }
 
     /// Clear the definition messages used to decode data messages and reset the CRC value. This
@@ -88,19 +105,23 @@ impl Deserializer {
         self.position += header.header_size() as usize;
         self.crc = 0;
 
-        if let Some(value) = header.crc() {
+        // Check CRC if the header was 14 bytes. If the value is 0 treat it like the CRC doesn't
+        // exist. This behavior doesn't appear to be documented but was verified using the
+        // FitTestTool.jar utility included with the SDK.
+        let crc_value = header.crc().unwrap_or(0);
+        if crc_value > 0 {
             let checksum = caculate_crc(&input[0..(header.header_size() - 2) as usize]);
-            if checksum != value {
-                warn!("invalid checksum header {} {}", checksum, value);
-                // return Err(Box::new(ErrorKind::InvalidCrc((
-                //     Vec::from(remaining),
-                //     FitObject::Header(header),
-                //     value,
-                //     checksum,
-                // ))));
-            }
-            if value == 0 {
-                self.crc = update_crc(0, &input[0..(header.header_size() as usize)]);
+            if !self
+                .options
+                .contains(&DecodeOption::SkipHeaderCrcValidation)
+                && checksum != crc_value
+            {
+                return Err(Box::new(ErrorKind::InvalidCrc((
+                    Vec::from(remaining),
+                    FitObject::Header(header),
+                    crc_value,
+                    checksum,
+                ))));
             }
         } else {
             // if the header doesn't have its own CRC then the header bytes are included in
@@ -115,14 +136,13 @@ impl Deserializer {
     fn deserialize_crc<'de>(&mut self, input: &'de [u8]) -> Result<(&'de [u8], FitObject)> {
         let (input, crc) = le_u16(input).map_err(|e| self.to_parse_err(e))?;
         self.position += 2;
-        if crc != self.crc {
-            warn!("invalid checksum crc {} {}", crc, self.crc);
-            // return Err(Box::new(ErrorKind::InvalidCrc((
-            //     Vec::from(input),
-            //     FitObject::Crc(crc),
-            //     crc,
-            //     self.crc,
-            // ))));
+        if !self.options.contains(&DecodeOption::SkipDataCrcValidation) && crc != self.crc {
+            return Err(Box::new(ErrorKind::InvalidCrc((
+                Vec::from(input),
+                FitObject::Crc(crc),
+                crc,
+                self.crc,
+            ))));
         }
         Ok((input, FitObject::Crc(crc)))
     }
@@ -159,8 +179,12 @@ impl Deserializer {
     /// is not easy to get using the vanilla From trait since we need outside information.
     fn to_parse_err(&self, err: nom::Err<nom::error::Error<&[u8]>>) -> crate::Error {
         match err {
-            nom::Err::Error(inner_err) => ErrorKind::ParseError(self.position, inner_err.code).into(),
-            nom::Err::Failure(inner_err) => ErrorKind::ParseError(self.position, inner_err.code).into(),
+            nom::Err::Error(inner_err) => {
+                ErrorKind::ParseError(self.position, inner_err.code).into()
+            }
+            nom::Err::Failure(inner_err) => {
+                ErrorKind::ParseError(self.position, inner_err.code).into()
+            }
             nom::Err::Incomplete(needed) => ErrorKind::UnexpectedEof(needed).into(),
         }
     }
@@ -187,6 +211,16 @@ impl FitStreamProcessor {
         Self::default()
     }
 
+    /// Add a decoding option to the processor
+    pub fn add_option(&mut self, opt: DecodeOption) {
+        self.deserializer.options_mut().insert(opt);
+    }
+
+    /// Add a decoding option to the processor
+    pub fn remove_option(&mut self, opt: DecodeOption) {
+        self.deserializer.options_mut().remove(&opt);
+    }
+
     /// Reset the decoder state and definition messages in use, this should be called at the end of
     /// each FIT file to ensure the accumlator fields in the decoder will produce the right values
     /// per file.
@@ -206,11 +240,16 @@ impl FitStreamProcessor {
     }
 }
 
-/// Deserialize a FIT file stored as an array of bytes and return the decoded data messages.
-pub fn from_bytes(mut buffer: &[u8]) -> Result<Vec<FitDataRecord>> {
+/// Deserialize a FIT file stored as an array of bytes and return the decoded data messages,
+/// with additional decode options
+pub fn from_bytes_with_options(
+    mut buffer: &[u8],
+    options: &HashSet<DecodeOption>,
+) -> Result<Vec<FitDataRecord>> {
     let mut processor = FitStreamProcessor::new();
     let mut records = Vec::new();
 
+    options.iter().for_each(|o| processor.add_option(*o));
     while !buffer.is_empty() {
         let (buf, obj) = processor.deserialize_next(buffer)?;
         match obj {
@@ -225,9 +264,22 @@ pub fn from_bytes(mut buffer: &[u8]) -> Result<Vec<FitDataRecord>> {
     Ok(records)
 }
 
-/// Deserialize a FIT file stored in a source that implements io::Read.
-pub fn from_reader<T: Read>(source: &mut T) -> Result<Vec<FitDataRecord>> {
+/// Deserialize a FIT file stored as an array of bytes and return the decoded data messages.
+pub fn from_bytes(mut buffer: &[u8]) -> Result<Vec<FitDataRecord>> {
+    from_bytes_with_options(&mut buffer, &HashSet::new())
+}
+
+/// Deserialize a FIT file stored in a source that implements io::Read, with additional decode options
+pub fn from_reader_with_options<T: Read>(
+    source: &mut T,
+    options: &HashSet<DecodeOption>,
+) -> Result<Vec<FitDataRecord>> {
     let mut buffer = Vec::new();
     source.read_to_end(&mut buffer)?;
-    from_bytes(&buffer)
+    from_bytes_with_options(&buffer, &options)
+}
+
+/// Deserialize a FIT file stored in a source that implements io::Read.
+pub fn from_reader<T: Read>(source: &mut T) -> Result<Vec<FitDataRecord>> {
+    from_reader_with_options(source, &HashSet::new())
 }