+++ src/profile/messages.rs
@@ -1,11 +1,12 @@
-//! Auto generated profile messages from FIT SDK Release: 21.47.00
+//! Auto generated profile messages from FIT SDK Release: 21.60.00
 #![allow(missing_docs)]
 #![allow(clippy::redundant_field_names)]
 #![allow(clippy::unreadable_literal)]
 use super::field_types::*;
 use super::{ComponentFieldInfo, FieldDataType, FieldInfo, MessageInfo};
 use std::collections::HashMap;
-pub const VERSION: &str = "21.47.00";
+pub const VERSION: &str = "21.60.00";
+/// Must be first message in file.
 pub fn file_id_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {
@@ -5020,6 +5021,42 @@ pub fn session_message() -> MessageInfo {
         components: Vec::new(),
     };
     fields.insert(200, field);
+    let field = FieldInfo {
+        name: "avg_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 208,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(208, field);
+    let field = FieldInfo {
+        name: "min_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 209,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(209, field);
+    let field = FieldInfo {
+        name: "max_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 210,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(210, field);
     // Sesson end time.
     let field = FieldInfo {
         name: "timestamp",
@@ -6574,6 +6611,42 @@ pub fn lap_message() -> MessageInfo {
         components: Vec::new(),
     };
     fields.insert(157, field);
+    let field = FieldInfo {
+        name: "avg_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 158,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(158, field);
+    let field = FieldInfo {
+        name: "min_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 159,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(159, field);
+    let field = FieldInfo {
+        name: "max_core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 160,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(160, field);
     // Lap end time.
     let field = FieldInfo {
         name: "timestamp",
@@ -7780,6 +7853,18 @@ m",
     };
     fields.insert(120, field);
     let field = FieldInfo {
+        name: "core_temperature",
+        field_type: FieldDataType::UInt16,
+        def_number: 139,
+        scale: 100.000000,
+        offset: 0.000000,
+        units: "C",
+        accumulate: false,
+        subfields: Vec::new(),
+        components: Vec::new(),
+    };
+    fields.insert(139, field);
+    let field = FieldInfo {
         name: "timestamp",
         field_type: FieldDataType::DateTime,
         def_number: 253,
@@ -8696,6 +8781,7 @@ pub fn device_info_message() -> MessageInfo {
         fields,
     }
 }
+/// Corresponds to file_id of workout or course.
 pub fn training_file_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {
@@ -8837,6 +8923,7 @@ pub fn training_file_message() -> MessageInfo {
         fields,
     }
 }
+/// Heart rate variability
 pub fn hrv_message() -> MessageInfo {
     let mut fields = HashMap::new();
     // Time between beats
@@ -11095,6 +11182,7 @@ pub fn course_point_message() -> MessageInfo {
         fields,
     }
 }
+/// Unique Identification data for a segment file
 pub fn segment_id_message() -> MessageInfo {
     let mut fields = HashMap::new();
     // Friendly name assigned to segment
@@ -11220,6 +11308,7 @@ pub fn segment_id_message() -> MessageInfo {
         fields,
     }
 }
+/// Unique Identification data for an individual segment leader within a segment file
 pub fn segment_leaderboard_entry_message() -> MessageInfo {
     let mut fields = HashMap::new();
     // Friendly name assigned to leader
@@ -11318,6 +11407,7 @@ pub fn segment_leaderboard_entry_message() -> MessageInfo {
         fields,
     }
 }
+/// Navigation and race evaluation point for a segment decribing a point along the segment path and time it took each segment leader to reach that point
 pub fn segment_point_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {
@@ -12558,6 +12648,7 @@ pub fn segment_lap_message() -> MessageInfo {
         fields,
     }
 }
+/// Summary of the unique segment and leaderboard information associated with a segment file. This message is used to compile a segment list file describing all segment files on a device. The segment list file is used when refreshing the contents of a segment file with the latest available leaderboard information.
 pub fn segment_file_message() -> MessageInfo {
     let mut fields = HashMap::new();
     // UUID of the segment file
@@ -14902,6 +14993,7 @@ pub fn hr_message() -> MessageInfo {
         fields,
     }
 }
+/// Value from 1 to 100 calculated by FirstBeat
 pub fn stress_level_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {
@@ -15702,6 +15794,7 @@ pub fn exd_data_concept_configuration_message() -> MessageInfo {
         fields,
     }
 }
+/// Must be logged before developer field is used
 pub fn field_description_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {
@@ -15878,6 +15971,7 @@ pub fn field_description_message() -> MessageInfo {
         fields,
     }
 }
+/// Must be logged before field description
 pub fn developer_data_id_message() -> MessageInfo {
     let mut fields = HashMap::new();
     let field = FieldInfo {